requirements.txt
# MacroHard - Security Analysis Platform
# Core dependencies

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# Security
cryptography==42.0.0
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# Analysis tools (installed separately in Docker for better layer caching)
# bandit==1.7.7
# semgrep (installed via binary in Dockerfile)
